% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg.R
\name{dg}
\alias{dg}
\title{Density function of the g distribution}
\usage{
dg(x, g)
}
\arguments{
\item{x}{real: numerical vector: values where the density has to be evaluated.}

\item{g}{positive real: shape (skewness) parameter.}
}
\value{
value of the density function of the g distribution.
}
\description{
This function evaluates the density of the g distribution
}
\examples{
yd <- dg(1,0.5)
}
