% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgh.R
\name{dgh}
\alias{dgh}
\title{Density function of the a g-and-h distribution}
\usage{
dgh(x, a, b, g, h, logar)
}
\arguments{
\item{x}{real: values where the density has to be evaluated.}

\item{a}{real: location parameter.}

\item{b}{positive real: scale parameter.}

\item{g}{real: first shape (skewness) parameter.}

\item{h}{positive real: second (heavy tailedness) shape parameter.}

\item{logar}{logical: if logar=0, the density is computed, if logar=1, the
log-density is computed.}
}
\value{
value of the density function of the g-and-h distribution.
}
\description{
This function computes numerically the density of a g-and-h distribution.
}
\examples{
yd <- dgh(1,0,1,0.5,.3,0)
}
