% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dghmix.R
\name{dghmix}
\alias{dghmix}
\title{Density function of a two-population mixture of g-and-h distributions}
\usage{
dghmix(x, p, a1, b1, g1, h1, a2, b2, g2, h2)
}
\arguments{
\item{x}{real: values where the density has to be evaluated.}

\item{a1}{real: location parameter of the first population.}

\item{b1}{positive real: scale parameter of the first population.}

\item{g1}{real: first shape (skewness) parameter of the first population.}

\item{h1}{positive real: second (heavy tailedness) shape parameter of the first population.}

\item{a2}{real: location parameter of the second population.}

\item{b2}{positive real: scale parameter of the second population.}

\item{g2}{real: first shape (skewness) parameter of the second population.}

\item{h2}{positive real: second (heavy tailedness) shape parameter of the second population.}
}
\value{
value of the density function of a mixture of two g-and-h distributions.
}
\description{
This function computes numerically the density of a mixture of two g-and-h distributions.
}
\examples{
yd <- dghmix(1,0,-1,0.5,.3,.2,1,1,.5,.1)
}
