% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lld.R
\name{lld}
\alias{lld}
\title{Log-likelihood function of the g distribution}
\usage{
lld(g, x)
}
\arguments{
\item{g}{real: shape (skewness) parameter.}

\item{x}{numerical vector: observations.}
}
\value{
value of the log-likelihood function of the g distribution.
}
\description{
This function evaluates the log-likelihood function of the g distribution
}
\examples{
yd <- lld(1,0.5)
}
