% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llgh.R
\name{llgh}
\alias{llgh}
\title{Log-likelihood function of the g-and-h distribution}
\usage{
llgh(x, y)
}
\arguments{
\item{x}{bivariate vector: values of the g and h parameters.}

\item{y}{numerical vector: observations.}
}
\value{
value of the log-likelihood function of the g distribution.
}
\description{
This function evaluates the log-likelihood function of the standard (a=0, b=1)
g-and-h distribution.
}
\examples{
yd <- lld(1,0.5)
}
