% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEBoot.R
\name{MLEBoot}
\alias{MLEBoot}
\title{MLE on samples obtained by resampling with replacement the original data.}
\usage{
MLEBoot(x, dati, nome)
}
\arguments{
\item{dati}{(n x 1) vector: observed  sample.}

\item{nome}{string: name of the file where the output is saved.}
}
\value{
Nothing. The matrix containing the estimated parameters at each
bootstrap replication are saved to fule pars.txt.
}
\description{
This function computes the MLEs of the parameters of #' a g-and-h distribution
on samples obtained by resampling with replacement the original data.
The output is to be used for computing standard errors via non-parametric
bootstrap.
}
