% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgh.R
\name{pgh}
\alias{pgh}
\title{Cumulative distribution function of the a g-and-h distribution}
\usage{
pgh(x, a, b, g, h)
}
\arguments{
\item{x}{real vector: values where the cdf has to be evaluated.}

\item{a}{real: location parameter.}

\item{b}{positive real: scale parameter.}

\item{g}{real: first shape (skewness) parameter.}

\item{h}{positive real: second (heavy tailedness) shape parameter.}
}
\value{
value of the cdf of the g-and-h distribution.
}
\description{
This function computes the cdf of a g-and-h distribution for a vectorial input.
}
\examples{
yd <- pgh(1,0,1,0.5,.3,0)
}
