% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qEst.R
\name{qEst}
\alias{qEst}
\title{Quantile estimation of the g-and-h distribution}
\usage{
qEst(x)
}
\arguments{
\item{x}{sample of observations from a g-and-h distribution.}
}
\value{
quantile estimates of the parameters of the g-and-h distribution.
}
\description{
This function the quantile estimators of the g-and-h distribution
}
\examples{
ySim <- rgh(100,1,1,0.5,0.3)
qPars <- qEst(ySim)
}
