% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgh.R
\name{qgh}
\alias{qgh}
\title{Quantile function of the a g-and-h distribution}
\usage{
qgh(p, a, b, g, h)
}
\arguments{
\item{a}{real: location parameter.}

\item{b}{positive real: scale parameter.}

\item{g}{real: first shape (skewness) parameter.}

\item{h}{positive real: second (heavy tailedness) shape parameter.}

\item{x}{real, 0<=x<=1: quantile levels.}
}
\value{
value of the quantile function of the g-and-h distribution.
}
\description{
This function computes the quantile function of a g-and-h distribution.
}
\examples{
yd <- qgh(.1,0,1,0.5,.3)
}
