% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qgh_deriv.R
\name{Qgh_deriv}
\alias{Qgh_deriv}
\title{First derivative of the g-and-h distribution quantile function}
\usage{
Qgh_deriv(x, b, g, h)
}
\arguments{
\item{x}{real: numerical vector: values where the function has to be evaluated.}

\item{b}{positive real: scale parameter.}

\item{g}{real: first shape (skewness) parameter.}

\item{h}{positive real: second (heavy tailedness) shape parameter.}
}
\value{
value of the first derivative of the g-and-h distribution
quantile function.
}
\description{
This function evaluates the first derivative of the g-and-h distribution
quantile function
}
\examples{
yf <- Qgh_deriv(1, 1, 0.3, 0.2)
}
