% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgh.R
\name{rgh}
\alias{rgh}
\title{Random number simulation from a g-and-h distribution}
\usage{
rgh(n, a, b, g, h)
}
\arguments{
\item{n}{positive integer: number of simulated random numbers.}

\item{a}{real: location parameter.}

\item{b}{positive real: scale parameter.}

\item{g}{real: first shape (skewness) parameter.}

\item{h}{positive real: second (heavy tailedness) shape parameter.}
}
\value{
n iid random numbers from the g-and-h distribution.
}
\description{
This function simulates random numbers from a g-and-h distribution
using its stochastic representation
}
\examples{
ySim <- rgh(100,1,1,0.5,0.3)
}
