% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rghmix.R
\name{rghmix}
\alias{rghmix}
\title{Simulation of a two-population mixture of g-and-h distributions}
\usage{
rghmix(n, p1, a1, b1, g1, h1, a2, b2, g2, h2)
}
\arguments{
\item{n}{real: number of random numbers simulated form the distribution.}

\item{a1}{real: location parameter of the first population.}

\item{b1}{positive real: scale parameter of the first population.}

\item{g1}{real: first shape (skewness) parameter of the first population.}

\item{h1}{positive real: second (heavy tailedness) shape parameter of the first population.}

\item{a2}{real: location parameter of the second population.}

\item{b2}{positive real: scale parameter of the second population.}

\item{g2}{real: first shape (skewness) parameter of the second population.}

\item{h2}{positive real: second (heavy tailedness) shape parameter of the second population.}
}
\value{
a sample of size n simulated from a mixture of two g-and-h distributions.
}
\description{
This function simulates a mixture of two g-and-h distributions.
}
\examples{
yd <- rghmix(100,0,-1,0.5,.3,.2,1,1,.5,.1)
}
