% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg.R
\name{dg}
\alias{dg}
\title{Density function of the g distribution}
\usage{
dg(x0, a, b, g)
}
\arguments{
\item{a}{real: location parameter.}

\item{b}{positive real: scale parameter.}

\item{g}{real: shape (skewness) parameter.}

\item{x}{real: numerical vector: values where the density has to be evaluated.}
}
\value{
Value of the density function of the g distribution.
}
\description{
dg evaluates the density of the g distribution
}
\examples{
yd <- dg(1,0,1,0.5)
}
